backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pytesseract==0.3.10
pillow>=10.2.0
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.4
openpyxl==3.1.2
aiofiles==23.2.1
bcrypt==4.1.2
PyJWT==2.8.0
# psycopg2-binary : optionnel, uniquement si vous utilisez PostgreSQL (pip install psycopg2-binary)
